# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Health"]
      }
    },
    "/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Health"]
      }
    },
    "/api/v1/ingest/signals": {
      "post": {
        "operationId": "IngestController_ingest",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "x-delivery-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "x-provider-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "202": { "description": "" } },
        "tags": ["Ingest"]
      }
    },
    "/api/v1/issues": {
      "get": {
        "operationId": "IssuesController_list",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}": {
      "get": {
        "operationId": "IssuesController_getOne",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}/evidence": {
      "get": {
        "operationId": "IssuesController_listEvidence",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}/audit": {
      "get": {
        "operationId": "IssuesController_listAudit",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}/timeline": {
      "get": {
        "operationId": "IssuesController_getTimeline",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}/evidence-graph": {
      "get": {
        "operationId": "IssuesController_getEvidenceGraph",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Issues"]
      }
    },
    "/api/v1/issues/{issueId}/close": {
      "post": {
        "operationId": "IssuesController_close",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "x-actor-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The resolved issue. A close that changed nothing (already resolved, or another close won the race) is also a 200 and is marked with `Idempotent-Replay: true`; its actor and reason were not recorded. 400: missing/invalid Idempotency-Key (uuid), X-Actor-Id (1-128) or reason (1-1000). 404: unknown issue or another tenant. 409: merged or removed.",
            "headers": {
              "Idempotent-Replay": {
                "description": "Present (`true`) only when this request closed nothing",
                "schema": { "type": "string", "enum": ["true"] }
              }
            }
          }
        },
        "tags": ["Issues"]
      }
    },
    "/api/v1/policy/rulesets": {
      "get": {
        "operationId": "PolicyRulesetsController_list",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyRulesets"]
      },
      "post": {
        "operationId": "PolicyRulesetsController_publish",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "x-actor-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["PolicyRulesets"]
      }
    },
    "/api/v1/policy/rulesets/{version}": {
      "get": {
        "operationId": "PolicyRulesetsController_getOne",
        "parameters": [
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyRulesets"]
      }
    },
    "/api/v1/policy/decisions": {
      "get": {
        "operationId": "PolicyDecisionsController_list",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyDecisions"]
      }
    },
    "/api/v1/policy/decisions/{decisionId}": {
      "get": {
        "operationId": "PolicyDecisionsController_getOne",
        "parameters": [
          {
            "name": "decisionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyDecisions"]
      }
    },
    "/api/v1/policy/decisions/{decisionId}/replay": {
      "post": {
        "operationId": "PolicyDecisionsController_replay",
        "parameters": [
          {
            "name": "decisionId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyDecisions"]
      }
    },
    "/api/v1/policy/dry-run": {
      "post": {
        "operationId": "PolicyEvaluationController_dryRun",
        "parameters": [
          {
            "name": "x-tenant-id",
            "required": true,
            "in": "header",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyEvaluation"]
      }
    },
    "/api/v1/policy/actions": {
      "get": {
        "operationId": "PolicyEvaluationController_list",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["PolicyEvaluation"]
      }
    }
  },
  "info": {
    "title": "Healer API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": { "schemas": {} }
}